split up navbar
